+++ dms_tools2/plot.py
@@ -768,8 +768,7 @@ def plotSiteDiffSel(names, diffselfiles, plotfile,
     y_lim = diffsel['diffsel'].max() #get max value used to plot the overlay
     # natural sort by site: https://stackoverflow.com/a/29582718
     diffsel = diffsel.reindex(index=natsort.order_by_index(
-            diffsel.index, natsort.index_natsorted(diffsel.site,
-            signed=True)))
+            diffsel.index, natsort.index_realsorted(diffsel.site)))
     # now some manipulations to make site str while siteindex is int
     diffsel['site'] = diffsel['site'].apply(str)
     diffsel['siteindex'] = pandas.Categorical(diffsel['site'],