+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Change Log
 ===========
 
-2.4.dev5
+2.4.5
 ----------
 * Handle single sample groups in `dms2_batch_diffsel` and `dms2_batch_fracsurvive`